Joshua_2018-10-24_0416
Fixed the conflict between two buttons in a form (it triggers submit & nested form is not allowed in HTML); changed button to an <a> tag instead. Hence <a> tag as a replacement, it can only be hidden instead of being "disabled" like a button. Also reverted some redundant redirects from footer.blade.php.